Add failing test case regarding nested `has_many` associations
In the current state of affairs, if we were to have factories building
nested `has_many` relationships the built models cannot be saved. When
`creating` a model with such a factory (like the shipment factory in
`ecto_has_many_test.exs`) a user would see the following error:

```
** (ArgumentError) model ExMachina.EctoHasManyTest.Package has value `[%ExMachina.EctoHasManyTest.PackageStatus{__meta__: #Ecto.Schema.Metadata<:built>, id: nil, package: #Ecto.Association.NotLoaded<association :package is not loaded>, package_id: nil, status: "ordered"}, %ExMachina.EctoHasManyTest.PackageStatus{__meta__: #Ecto.Schema.Metadata<:built>, id: nil, package: #Ecto.Association.NotLoaded<association :package is not loaded>, package_id: nil, status: "sent"}, %ExMachina.EctoHasManyTest.PackageStatus{__meta__: #Ecto.Schema.Metadata<:built>, id: nil, package: #Ecto.Association.NotLoaded<association :package is not loaded>, package_id: nil, status: "shipped"}]` set for assoc named `statuses`. Assocs can only be manipulated via changesets, be it on insert, update or delete.
```

This happens because Ecto's `put_assoc` won't allow `built` associations
in its `value`. This will be solved in a future commit.