Add name_from_struct for strategies (#116)
Options passed to handling functions are now a map which makes it easier
to pattern match. It also includes the calling factory module in the
options.

name_from_struct/1 is added so that people can create strategies that
define callbacks or need to call other functions with the factory name.

Closes #111. See discussion for more details on why this strategy was
chosen.