Don't return id field in params_for
I can't think of any use case where you would want the id field when
using params_for. For something scenarios like submitting to an API, or
getting the bare fields to test validations it would be rare to
actually submit the id itself, and so the current behavior feels
unexpected.