Clarify where to import factory module in Phoenix
Being new to Phoenix I spent some time this morning trying to figure out why I got the error `undefined function insert/1` when setting up ExMachina. Turns out I'd put the `import MyApp.Factory` into the body of MyApp.ConnCase instead of in the `using` block. 

This PR adds a note to the readme in the hope of saving other developers the time it took to figure that out.

Thanks for the great library!