Support Ecto 2.2.0-rc.0
These changes allow for the the library to work with changes presented in the current state of Ecto 2.2.0-rc.0.

Adds additional match for `{name, source, type}` supporting the new `source` option on field.

Updates test for persisting non-schema attributes as new warning is emitted when passing a different schema type to relation.