Support structs in string_params_for/2 (#245)
`string_params_for/2` currently fails when a struct is passed for a
field. The error comes when a function from the enumerable protocol is
called on the struct (assuming it's a plain map). So the following
example would fail,

```elixir
Factory.string_params_for(:model, some_field: %SomeStruct{key: value})
```

The behavior of `string_params_for/2` with a struct should be to convert
it into a plain map with strings as keys. So in the example above, we
would expect the result from `string_params_for/2` to include
`%{"some_field" => %{"key" => value}}`

This commit does this by using `Map.from_struct/1` to turn the struct
into plain map before continuing with the process of converting the
atoms into strings.