Handle nil values for has_one associations (#266)
The has_one association allows for the referenced value to be nil.

ExMachina by default sets such fields to `%Ecto.Association.NotLoaded{}` when the field Is not mentioned in the factory definition. However, setting the field to nil explicitly in the factory definition caused a CaseClauseError in ExMachina.Ecto.handle_assoc.