Remove unused function (#287)
Fixes #157

Since version 1.0.0, when users of ex_machina try to call the deprecated fields_for/2 function, it raises an exception. This doesn't seem to provide any benefit. If instead it's removed, calls to it will fail at compile time. But if it stays in, calls will fail at runtime, possibly in production.

This PR solves the Dialyzer issue and it provides a better solution for dealing with the obsolete function.