Add additional callbacks for functions with default params
In pretty much every public function in `ex_machina` you allow users to
either give or not give parameters for the resource that's being built,
but in the documentation that's not shown because you can't show
optional arguments in callback definitions.

This commit adds callbacks for these `/1` or `/2` functions that are
commonly used. This benefits documentation, but also enforces the
behaviour contract to make sure that all documented functionality is
implemented correctly.