Allow more control over factory definitions
Description
-----------

Currently `build(:user)` will grab the struct/map defined in
`user_factory/0` and merge any attributes passed into it via `struct!/2`
or `Map.merge/2`.

It would be nice to allow users to have more control of factories if
they want to, while maintaining the ease of use for the majority of
cases. It seems this can be done by allowing users to change the
`user_factory/0` to a factory that takes in the attributes being passed
in.

We can do so by checking if they have defined a factory with `/1` arity
(meaning they allow parameters to be passed).

Example
---------

Let's take a look at a scenario where this might be wanted:

```elixir
def user_factory do
  name = "John"
  email = sequence(:email, &"#{name}-#{&1}@foo.com")

  %User{
    name: name,
    email: email,
    age: 20
  }
end

build(:user)

build(:user, name: "James")
```

As we can see the issue is that since we derive the `email` from the
`name`, the email has the old name in it, even when we pass "James" as
the name. It would be nice to be able to get the `name` from the
attributes _before_ we create the email.

So we could define this instead,

```elixir
def user_factory(attrs) do
  name = Map.get(attrs, :name, "John")
  email = sequence(:email, &"#{name}-#{&1}@foo.com")

  user = %User{
    name: name,
    email: email,
    age: 20
  }

  merge_attributes(user, attrs)
end

build(:user)

build(:user, name: "James", age: 30)
```

Note the inclusion of `merge_attributes/2` in our example above. We need
that so that `age` gets properly overridden from `20` to `30`. Without
it, the only attributes that will be set are `name` and `email`.

The `merge_attributes/2` function is a new function we expose to make it
easier for users to merge attributes regardless of whether it is a
struct or a map. It essentially delegates to `struct!/2` or
`Map.merge/2`.