Format elixir code
We use `mix format` to format all our elixir code.

Since we want to check that the code is formatted, we update the docker
image in circle CI to elixir 1.7.4.

And since we are now setting a version in CI, we also add a
`.tool-versions` with elixir 1.7.4 and erlang 21 so that contributors
can use consistent versions.