Allow ExMachina.Ecto to be used without the :repo option (#370)
We should be able to use Ecto without a repo. This is useful when using ecto without ecto_sql.

ExMachina currently raising an ArgumentError when using ExMachina.Ecto without a repo. This change now allows the repo to be nil and raises a RuntimeError when the insert/1 function is called instead.

New error message should still be helpful enough for people that simply forgot to add the :repo option.

```
insert/1 is not available unless you provide the :repo option. Example:

use ExMachina.Ecto, repo: MyApp.Repo
```