Add build_lazy/2
We add a potential solution to lazily build factories.

The problem
------------

People do this:

```elixir
build_pair(:account, email: build(:email))
```

They think that each account will have a different email (especially if
they set a sequence). The problem is that this is just function calling.
So the above is equivalent to this:

```elixir
email = build(:email)
build_pair(:account, email: email)
```

In other words, people are trying to insert multiple accounts with the
same email. They'll get a validation error if they have uniqueness.

Solution
--------

The solution is to somehow delay evaluation of the factory. The simplest
way to do this is to capture it in an anonymous function, basically
doing this:

```elixilr
build_pair(:account, email: fn -> build(:email) end)
```

A more advanced implementation might create a struct (a type) that is
executed later.

That's what we aim to do with `build_lazy/2` (function name still in
question). `build_lazy/2` literally returns `fn -> build(:name) end`.

The trick lies in the `build/2` function. It now checks all
attributes passed to it. If it's a function, it evaluates it. That way,
when we're building a list of things, we evaluate the `build(:email)`
once per account we're creating.