Spike building ExMachina.Factory structs
What changed?
=============

We spike building lazy factories with a struct rather than a function.
`build_lazy/2` returns an `%ExMachina.Factory{}` struct which contains
all the parameters needed to build the factory.

Why?
----

A previous spike used an anonymous function to build lazy factories. But
I do not know if that would be a breaking change. Some people could
conceivably define a factory with a captured function as one of its
attributes. If they do that, and we ship build_lazy with anonymous
functions, then we'd accidentally execute the developer's factory
because we only check if it's a function.

By relying on an internal struct (`%ExMachina.Factory{}`), we can
pattern match on the type (being the struct) and build it lazily without
accidentally evaluation functions that a developer may have set as
factory attribute.