Introduce `build_lazy/2`
The problem
------------

People want a different email per account (especially if the email
factory has a sequence), but they do this:

```elixir
build_pair(:account, email: build(:email))
```

The problem is that `build/2` is just function calling. So the above is
equivalent to this:

```elixir
email = build(:email)
build_pair(:account, email: email)
```

In other words, we get on email factory for all of the accounts.

The problem is made worse when using it with Ecto. We can imagine the
following scenario:

```elixir
insert_pair(:account, user: build(:user))
```

If the user factory has a uniqueness constraint, `insert_pair/2` will
raise an error because we'll try to insert a user with the same value
(even if using a sequence).

Solution
--------

The solution is to delay evaluation of the factory. We do this by
introducing a private struct `%ExMachina.Instance{}` to hold the data
necessary to build the instance of that factory.

Note that this struct is private and liable to change, so users of the
library shouldn't depend on it.

The trick then lies in the `build/2` function. We make it a terminal
function in that it will evaluate any lazy factories recursively. To do
that, we update the `build/2` function to evaluate `ExMachina.Instance`
structs when iterating through its attributes (even when an attribute
has a list of factories).