Move lazy factory evaluation after factory creation
We move the evaluation of lazy factories until after the creation of the
factory. This allows us to use `build_lazy` from within factory
definitions and not just when combined with `build/2` or `build_*`
functions.

This also opens the possibility of passing the built factory as an
argument to build_lazy but we'd have to modify how it works:

E.g.

```elixir
def account_factory do
  %{
    private: true,
    profile: build_lazy(fn account ->
      build(:profile, private: account.private
    end)
  }
end
```