Rename Instance.build -> InstanceTemplate.evaluate
We rename Instance -> InstanceTemplate & build -> evaluate.

We do that because it's really a template for an instance in the sense
that we have the factory's name and extra attributes, but any kind of
sequence in the factory definition will still auto increment, making the
evaluation of two instance templates not be the same.

In other words, given the factory:

```elixir
def user_factory do
  %{email: sequence("email")}
end
```

Then evaluating the same build lazy template will not give use the same
sequence:

```elixir
template = build_lazy(:user)

InstanceTemplate.evaluate(template) != InstanceTemplate.evaluate(template)
```