Evaluate lazy factories before full-control factories
We evaluate lazy factories before passing the attrs to the factories
that have full control. That means that we can still use `build_lazy`
within `build/2` and `build_*` functions that have a factory with full
control, but we cannot use `build_lazy/2` as part of the definition of a
factory that has full control.

i.e. we can do this:

```elixir
def account_factory(attrs) do
  %{
    user: build(:user),
    ...
  }
end

build(:account, build_lazy(:user))
```

But we cannot do this:

```elixir
def account_factory(attrs) do
  %{
    user: build_lazy(:user),
    ...
  }
end
```

I'm not sure if we should even do that, but it seems like a good
compromise so people can continue using factories with full control,
minus the ability to define lazy factories in the definition, since
ExMachina does nothing after they define it.