Allow setting sequence starting point
What changed?
============

We introduce `ExMachina.sequence/3` that allows for passing the
`start_at` option to set a starting point for a sequence.

This is meant to be used in a factory definition:

```elixir
def money_factory do
  %{
    cents: sequence(:cents, &"#{&1}", start_at: 60)
  }
end
```

Then, the following would increment starting at 60:

```elixir
build(:money)
// => %{cents: "60"}

build(:money)
// => %{cents: "61"}
```