feat: Add support for StreamData generators
This is a WIP PR that adds support for defining "generators", `StreamData` based
functions that allow users to create property tests (using `ExUnitProperties`),
and which overloads the existing `build` functions to use generators if they
exist.

The goal is to allow users to define generators like this:

```elixir
defmodule MyApp.Factory do
  # without Ecto
  use ExMachina

  def user_generator do
    gen all name <- do
      %MyApp.User{
        name: name
      }
    end
  end
end
```

And then, in tests, `build(:user)` would return the first element created from
that `StreamData` stream, while `generate(:user)` (a new function) would return
the stream itself, to be used in property tests. The benefit is that users
wouldn't have to both create a ExMachina factory for their regular unit tests
and a `StreamData` generator for their property tests.

Users may define both a generator and a factory (e.g. `user_generator` and
`user_factory`), in which case the factory will take precedence over the
generator when calling `build`, but the generator will still be used for
`generate`. Using lazy attributes in generators is not supported, because using
`StreamData` solves the same problem.

I am not 100% sure this is useful in it's current incarnation, how it should
play together with the `sequence` function or how to test this properly, but
some preliminary local testing using another application seemed to indicate that
the basic functionality works as expected.

As such, I am not looking to get this merged as is, but to start a conversation
about whether something like this is desirable in ExMachina and, if so, how we
might bring this PR (or another PR) to a state that could get it merged.

Inspiration: https://github.com/beam-community/ex_machina/issues/351 and
https://hexdocs.pm/ecto_stream_factory/readme.html